Apply s3utils.EncodePath to match signature gen
Don't rely on `net.URL` to escape, as it correctly applies RFC 3986 and doesn't bother escaping a `+` in the URL. Instead, apply `s3utils.EncodePath` to the URL, matching the way the minio signature package builds the canonical URL and avoiding a signature mismatch.

This change should be accompanied by reverting https://github.com/mattermost/mattermost/pull/23849, but we'll still need to keep the code to try the "newer" path until we can migrat the files incorrectly named during the deployment window.

Fixes: https://mattermost.atlassian.net/browse/MM-51753