Add request validation
This change adds optional http request validation logic which can
be enabled via the server config. This validation provides an
additional layer of security by performing a reverse address
lookup and comparing the name against a standard kubernetes
namespace check. As we use installation IDs as a direct mapping
to Mattermost workspaces, we can require that the ID of the request
and request namespace match.

This change also corrects an issue where installation IDs were not
being reported correctly for metrics.